Update doc to reference VimwikiListChangeLvl
Just started trying out vimwiki yesterday. Seems like the doc refers to `:VimwikiListChangeLevel` whereas the command itself is `:VimwikiListChangeLvl` and I don't see any hint of VimwikiListChangeLevel within the codebase.

Unless this was some rename that happened and an alias needs to be provided, seems like a quick update to the doc should do the trick?